+++ packages/pwa-kit-create-app/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## v3.18.0-dev (Mar 20, 2026)
+- Allow auth related cookies domain to be set via config [#3782](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3782)
 
 ## v3.17.1 (Mar 20, 2026)
 - Add base path prefix to support multiple MRT environments under 1 domain [#3614](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3614)